# openapi.yaml
openapi: 3.0.3
info:
  title: LoadFlow API
  version: 1.3.0
  description: |
    Welcome to the LoadFlow API!
    
    This API provides a complete backend for the LoadFlow Dispatch Application, allowing dispatchers, drivers, and administrators to orchestrate logistics seamlessly.

    ### 🔑 Authentication

    The system uses **Cookie-based JWT Authentication**.

    1. **Login** or **Register** (if invited) to receive a secure HTTP-only cookie.
    2. This cookie is automatically sent with subsequent requests.
    3. **Note for API Tools (Postman/Swagger):** Ensure cookies are enabled/stored after login so protected endpoints can be tested.

    ### 📦 Primary Features
    
    This API is built explicitly to handle fleet operations.
    
    - **Logistics Data:** Access resources in JSON format.
    - **Driver Dashboard Metrics:** Real-time calculation of Service Hours, Miles, and Completed Missions.
    - **Smart Notifications:** Optimistic state updates and ultra-fast polling for instant driver/dispatcher feedback.
    - **Proof of Delivery:** Secure `multipart/form-data` uploads handled and routed towards Cloudinary.
    - **Server Actions:** Public gateways directly dispatch SMTP NodeMailer notifications without persistence overhead.

    ### 👥 User Roles

    - **Driver:** Can view assigned configurations, update transit statuses, track performance metrics (Miles/Hours), and submit Proof of Delivery.
...
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/auth/check-email:
    get:
      tags: [Auth]
      summary: Check whether an email exists and whether it is still pending activation
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Email lookup result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckEmailResponse'
        '400':
          description: Missing email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin accounts cannot register through the public flow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Email not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset for admin approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Request accepted or suppressed for privacy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Missing email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/auth/2fa/generate:
    post:
      tags: [Auth]
      summary: Generate a 2FA secret and QR code for the logged-in user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: 2FA setup secret generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                  qrCode:
                    type: string
                    description: Data URL for the QR code image
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/2fa/verify:
    post:
      tags: [Auth]
      summary: Verify a 2FA code and enable 2FA for the logged-in user
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        '200':
          description: 2FA enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid request or code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/audit:
    get:
      tags: [Audit]
      summary: Fetch audit logs
      description: Admin only
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  total:
                    type: integer
                  page:
                    type: integer
                  totalPages:
                    type: integer
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    AuditLog:
      type: object
      properties:
        _id:
          type: string
        userId:
          oneOf:
            - type: string
              description: User ID as a string
            - type: object
              properties:
                _id:
                  type: string
                name:
                  type: string
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [Admin, Dispatcher, Driver]
          nullable: true
        action:
          type: string
        entityType:
          type: string
          nullable: true
        entityId:
          type: string
          nullable: true
        details:
          type: object
          additionalProperties: true
          nullable: true
        ipAddress:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    CheckEmailResponse:
      type: object
      properties:
        role:
          type: string
          enum: [Driver, Dispatcher]
        isPending:
          type: boolean
        name:
          type: string
...